autotest_framework/requirements.txt
pytest==7.4.0
requests==2.31.0
pytest-html==4.1.1
allure-pytest==2.13.0
jsonschema==4.19.0
jsonpath-ng==1.6.0
pyyaml==6.0.1
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
pytest-timeout==2.1.0
selenium==4.12.0
webdriver-manager==4.0.1
faker==19.6.2
python-dotenv==1.0.0
pydantic==2.3.0